Succesfully tested RestAPIs in swagger

+++ Create-Task-Application/src/main/java/com/batch5/Create_Task_Application/UserModule/exceptions/RoleNotFoundException.java
@@ -1,7 +1,7 @@
 package com.batch5.Create_Task_Application.UserModule.exceptions;
 
 public class RoleNotFoundException extends AppException {
-    public RoleNotFoundException(Long id) {
+    public RoleNotFoundException(Integer id) {
         super("Role not found with ID: " + id, 404);
     }
     public RoleNotFoundException(String roleName) {
